chore(scripts) Extract build part to separate script for local testing
