+++ package.json
@@ -32,6 +32,7 @@
     "storybook": "storybook dev -p 4000 --ci --disable-telemetry",
     "build-storybook": "storybook build",
     "lib:release": "node scripts/release.js --run",
+    "lib:build": "node scripts/build.js --run",
     "prepare": "husky"
   },
   "exports": {
